+++ lib/src/defines.cuh
@@ -5,7 +5,9 @@
 #ifndef OCTREECUDA_DEFINES_CUH
 #define OCTREECUDA_DEFINES_CUH
 
-constexpr unsigned int BLOCK_SIZE_MAX = 1024;
-constexpr unsigned int GRID_SIZE_MAX = 65535;
+#include <cstdint>
+
+constexpr uint64_t BLOCK_SIZE_MAX = 1024;
+constexpr uint64_t GRID_SIZE_MAX = 65535;
 
 #endif //OCTREECUDA_DEFINES_CUH
